# models/product.js
class Product {
  constructor(id, productName, productDescription, productPrice, productStock) {
    this.id = id;
    this.productName = productName;
    this.productDescription = productDescription;
    this.productPrice = productPrice;
    this.productStock = productStock;
  }
}

module.exports = Product;